requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.0
cheroot==8.6.0
ciso8601==2.2.0
click==8.0.3
#cloud-init==22.2
colorama==0.4.4
colored==1.4.3
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dateparser==1.1.1
dbus-python==1.2.18
distro==1.7.0
#distro-info===1.1build1
Flask==2.1.3
Flask-RESTful==0.3.9
frozenlist==1.3.0
gevent==21.12.0
gmqtt==0.6.11
greenlet==1.1.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jaraco.functools==3.5.1
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.1
more-itertools==8.10.0
multidict==6.0.2
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.3
pandas-ta==0.3.14b0
patsy==0.5.2
pexpect==4.8.0
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
#PyGObject==3.42.0
PyHamcrest==2.0.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
#python-apt==2.3.0+ubuntu2.1
python-binance==1.0.16
python-dateutil==2.8.2
#python-debian===0.1.43ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2022.3.2
requests==2.25.1
scipy==1.9.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
statsmodels==0.13.2
systemd-python==234
ta==0.10.1
TA-Lib==0.4.24
Twisted==22.1.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.9
ufw==0.36.1
ujson==5.4.0
unattended-upgrades==0.1
unicorn-binance-websocket-api==1.41.0
unicorn-fy==0.12.2
urllib3==1.26.5
wadllib==1.3.6
websocket==0.2.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.1
yarl==1.7.2
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
